+++ tests/UnitTestEx.NUnit.Test/Other/ExpectationsTest.cs
@@ -15,7 +15,7 @@ namespace UnitTestEx.NUnit.Test.Other
             var gt = GenericTester.Create().ExpectException().Any();
 
             var ex = Assert.Throws<Exception>(() => ArrangerAssert(async () => await gt.ExpectationsArranger.AssertAsync(null, null)));
-            Assert.AreEqual("Expected an exception; however, the execution was successful.", ex.Message);
+            Assert.That(ex.Message, Is.EqualTo("Expected an exception; however, the execution was successful."));
 
             ArrangerAssert(async () => await gt.ExpectationsArranger.AssertAsync(null, new DivideByZeroException()));
         }
@@ -28,7 +28,7 @@ namespace UnitTestEx.NUnit.Test.Other
             ArrangerAssert(async () => await gt.ExpectationsArranger.AssertAsync(null, new DivideByZeroException("Error")));
 
             var ex = Assert.Throws<Exception>(() => ArrangerAssert(async () => await gt.ExpectationsArranger.AssertAsync(null, new DivideByZeroException("not ok"))));
-            Assert.AreEqual("Expected Exception message 'error' is not contained within 'not ok'.", ex.Message);
+            Assert.That(ex.Message, Is.EqualTo("Expected Exception message 'error' is not contained within 'not ok'."));
         }
 
         [Test]
@@ -37,10 +37,10 @@ namespace UnitTestEx.NUnit.Test.Other
             var gt = GenericTester.Create().ExpectException().Type<DivideByZeroException>();
 
             var ex = Assert.Throws<Exception>(() => ArrangerAssert(async () => await gt.ExpectationsArranger.AssertAsync(null, null)));
-            Assert.AreEqual("Expected an exception; however, the execution was successful.", ex.Message);
+            Assert.That(ex.Message, Is.EqualTo("Expected an exception; however, the execution was successful."));
 
             ex = Assert.Throws<Exception>(() => ArrangerAssert(async () => await gt.ExpectationsArranger.AssertAsync(null, new NotSupportedException())));
-            Assert.AreEqual("Expected Exception type 'DivideByZeroException' not equal to actual 'NotSupportedException'.", ex.Message);
+            Assert.That(ex.Message, Is.EqualTo("Expected Exception type 'DivideByZeroException' not equal to actual 'NotSupportedException'."));
 
             ArrangerAssert(async () => await gt.ExpectationsArranger.AssertAsync(null, new DivideByZeroException()));
         }
@@ -51,7 +51,7 @@ namespace UnitTestEx.NUnit.Test.Other
             var gt = GenericTester.Create().ExpectError("No error will be raised.");
 
             var ex = Assert.Throws<Exception>(() => ArrangerAssert(async () => await gt.ExpectationsArranger.AssertAsync(null, null)));
-            Assert.AreEqual("Expected one or more errors; however, none were returned.", ex.Message);
+            Assert.That(ex.Message, Is.EqualTo("Expected one or more errors; however, none were returned."));
         }
 
         [Test]
@@ -61,10 +61,10 @@ namespace UnitTestEx.NUnit.Test.Other
             ArrangerAssert(async () => await gt.ExpectationsArranger.AssertValueAsync(null, "bob"));
 
             var ex = Assert.Throws<Exception>(() => ArrangerAssert(async () => await gt.ExpectationsArranger.AssertValueAsync(null, "jenny")));
-            Assert.IsTrue(ex.Message.Contains("Value is not equal: \"bob\" != \"jenny\"."));
+            Assert.That(ex.Message.Contains("Value is not equal: \"bob\" != \"jenny\"."), Is.True);
 
             ex = Assert.Throws<Exception>(() => ArrangerAssert(async () => await gt.ExpectationsArranger.AssertValueAsync(null, null)));
-            Assert.IsTrue(ex.Message.Contains("Kind is not equal: String != Null."));
+            Assert.That(ex.Message.Contains("Kind is not equal: String != Null."), Is.True);
         }
 
         [Test]
@@ -74,10 +74,10 @@ namespace UnitTestEx.NUnit.Test.Other
             ArrangerAssert(async () => await gt.ExpectationsArranger.AssertValueAsync(null, "bob"));
 
             var ex = Assert.Throws<Exception>(() => ArrangerAssert(async () => await gt.ExpectationsArranger.AssertValueAsync(null, "jenny")));
-            Assert.IsTrue(ex.Message.Contains("Value is not equal: \"bob\" != \"jenny\"."));
+            Assert.That(ex.Message.Contains("Value is not equal: \"bob\" != \"jenny\"."), Is.True);
 
             ex = Assert.Throws<Exception>(() => ArrangerAssert(async () => await gt.ExpectationsArranger.AssertValueAsync(null, null)));
-            Assert.IsTrue(ex.Message.Contains("Kind is not equal: String != Null."));
+            Assert.That(ex.Message.Contains("Kind is not equal: String != Null."), Is.True);
         }
 
         [Test]
@@ -89,7 +89,7 @@ namespace UnitTestEx.NUnit.Test.Other
             ArrangerAssert(async () => await gt.ExpectationsArranger.AssertValueAsync(null, new { Id = 88, Name = "bob" }));
 
             var ex = Assert.Throws<Exception>(() => ArrangerAssert(async () => await gt.ExpectationsArranger.AssertValueAsync(null, new { Id = 99, Name = "bob" })));
-            Assert.IsTrue(ex.Message.Contains("Path '$.id': Value is not equal: 88 != 99."));
+            Assert.That(ex.Message.Contains("Path '$.id': Value is not equal: 88 != 99."), Is.True);
 
             gt = GenericTester.CreateFor<Entity<int>>().ExpectValue(new Entity<int> { Id = 88, Name = "bob" }, "id");
             ArrangerAssert(async () => await gt.ExpectationsArranger.AssertValueAsync(null, new { Id = 99, Name = "bob" }));